requirements.txt
fastapi[all]
sqlalchemy
sqlalchemy_utils
PyJWT
pyjwt[crypto]
passlib
bcrypt